+++ Classes/MP4Tags.mm
@@ -41,6 +41,7 @@ using namespace std;
 		data->f->mp4 = new MP4::File([filename UTF8String]);
 		data->file = data->f->mp4;
 		[self initFields];
+		kind = @"MP4";
     }
     
     return self;